% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore a data set using shiny}
\usage{
explore(data)
}
\arguments{
\item{data}{A data frame or matrix}
}
\value{
Opens a shiny app
}
\description{
Explore a data set using shiny
}
\examples{
\dontrun{
explore(iris)
}

}
