% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholarCloud.R
\name{scholarCloud}
\alias{scholarCloud}
\title{Word clouds from Google Scholar IDs}
\usage{
scholarCloud(scholarId, minFreq = 1, maxWords = 100,
  colours = viridis::viridis(8))
}
\arguments{
\item{scholarId}{The ID number of person's Google scholar page. This is usually an alphanumeric code in the URL of their Google scholar page}

\item{minFreq}{The minimum frequency of words used in the word cloud}

\item{maxWords}{The maximum number of words to show on the page}

\item{colours}{The colour palette}
}
\value{
Returns a word cloud plot
}
\description{
Take the Google Scholar IDs of a person and create a word cloud of their paper titles
}
\examples{
\dontrun{
id = 'mXSv_1UAAAAJ' # Leo Brieman
scholarCloud(id)
}
}
