% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{packages}
\alias{packages}
\title{Function to load in packages and install those which are missing}
\usage{
packages(...)
}
\arguments{
\item{...}{A character vector of packages to be loaded/installed}
}
\value{
A list of the currently installed packages
}
\description{
Function to load in packages and install those which are missing
}
\examples{
\dontrun{
packages(c('Bchron', 'simmr'))
}
}
